Release 0.3.0: publication-ready polish
Publication-readiness pass: tkk was not in a state where I would be
comfortable pointing a reviewer at it even though the code itself is
solid. This commit fixes the packaging, metadata, documentation, and
CI gaps.

Fixed:
- eval.fusion.rrf tie-breaking: the perturbation was 1e-10 added to
  float32 scores, which rounds away (float32 eps ~= 1.19e-7). Every
  tied row therefore sorted identically and position 0 always won.
  Now done in float64. Restores the two previously-failing tests
  (test_rrf_seeded_reproducible, test_rrf_fair_tiebreaking).

Added:
- LICENSE (MIT)
- CHANGELOG.md with 0.1.0 / 0.2.0 / 0.3.0 history
- src/kge_kernels/py.typed marker (PEP 561)
- .github/workflows/tests.yml: pytest on py3.10/3.11/3.12 + ruff lint
- __version__ = "0.3.0" on the package

Changed:
- README.md rewritten to reflect the actual surface area (framework
  primitives, KGE models, losses, data utilities, ranking metrics,
  Evaluator, logging, sampler, adapter, partial scoring). Previous
  README described only sampler/scoring/partial and claimed the public
  API was "centered on corrupt and score" — that stopped being true
  several versions ago.
- pyproject.toml: version 0.2.0 -> 0.3.0. Added authors, urls,
  keywords, classifiers, license metadata, py.typed package-data.
  Description updated to match the real scope.
- Import order cleaned up across 15 files via ruff --fix. Removed
  two misplaced semicolon statements, one unused local, one
  module-level-import-after-code, and one lambda assignment that
  ruff's E731 flags.

Test suite: 153 passed, 0 failed (was 151 passed / 2 failed RRF).

+++ src/kge_kernels/eval/fusion.py
@@ -56,9 +56,14 @@ def rrf(
         # Reshape flat [CQ*K] → [CQ, K] (column-major pool layout)
         scores_2d = scores_flat.view(K, CQ).t()
 
-        # Rank with seeded noise for fair tie-breaking (1e-10 matches v1 behavior)
-        noise = torch.rand(CQ, K, generator=gen, device=device, dtype=scores_2d.dtype) * 1e-10
-        sorted_idx = (scores_2d + noise).argsort(dim=1, descending=True)
+        # Rank with seeded noise for fair tie-breaking. The addition must
+        # happen in float64: float32 has ~1.19e-7 precision, so a
+        # 1e-10 perturbation added to a float32 score is invisible and
+        # argsort's stable ordering would always favor position 0.
+        # float64 precision (~2.2e-16) keeps the perturbation observable
+        # while staying well below any realistic score difference.
+        noise = torch.rand(CQ, K, generator=gen, device=device, dtype=torch.float64) * 1e-10
+        sorted_idx = (scores_2d.to(torch.float64) + noise).argsort(dim=1, descending=True)
         ranks = torch.zeros_like(sorted_idx, dtype=torch.long)
         ranks.scatter_(1, sorted_idx, arange_k + 1)
